Cadent Aperture MX Bid Adapter: rename files to match bidder code
