+++ modules/cadent_aperture_mxBidAdapter.js
